requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
openai
python-dotenv
langsmith
tiktoken
ipykernel
notebook
beautifulsoup4   
faiss-cpu
streamlit
langchain-groq
groq
huggingface_hub
transformers
sse-starlette